requirements.txt
discord.py>=2.3.0
aiosqlite>=0.19.0
imagehash>=4.3.1
Pillow>=10.0.0
aiohttp>=3.9.0
python-dotenv>=1.0.0